pyproject.toml
[tool.poetry]
name = "lidl-receipt-client"
version = "0.1.0"
description = ""
authors = ["Mohamad Ramadan <dev@ramda.net>"]
readme = "README.md"

[tool.poetry.scripts]
fetch-receipts = 'lidl_receipt_client.main:fetch_all_receipts'
analyze = 'lidl_receipt_client.main:analyze_receipts'

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.28.2"
pyjwt = "^2.6.0"
dataclasses-json = "^0.5.7"
python-dateutil = "^2.8.2"
click = "^8.1.3"

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
